Optimize imports with big moving.
